chore: organize project detailspage components
