+++ src/app/(main)/projects/[id]/_components/export-pdf/index.ts
@@ -1,0 +1,1 @@
+export * from "./qcp-report";
