+++ src/app/(main)/projects/[id]/_components/index.ts
@@ -1,0 +1,6 @@
+export * from "./action-buttons";
+export * from "./export-pdf";
+export * from "./project-details";
+export * from "./project-details/project-details-skeleton";
+export * from "./table";
+export * from "./test-columns";
