+++ src/app/(main)/projects/[id]/_components/project-details/index.ts
@@ -1,0 +1,2 @@
+export * from "./contract-details";
+export * from "./project-info-button";
