+++ src/app/(main)/projects/[id]/_components/table/index.ts
@@ -1,0 +1,3 @@
+export * from "./materials-table";
+export * from "./table";
+export * from "./work-items-table";
