+++ src/app/(main)/projects/[id]/_components/test-columns/index.ts
@@ -1,0 +1,2 @@
+export * from "./test-counter";
+export * from "./test-status";
